refactor: rename x.interface.ts to x.interfaces.ts
idk just seems like better naming, usually contain multiple
