+++ server/src/modules/filestore/file-store.interfaces.ts
